MH3-112385: Regeneration SAMD20 with low in RAM and Flash

+++ apps/prime_apps/prime_service_1_4_modem_monolithic/src/config/samd20_xpro_pl460/service/pvddmon/srv_pvddmon.c
@@ -167,7 +167,7 @@ void SRV_PVDDMON_CallbackRegister (SRV_PVDDMON_CALLBACK callback, uintptr_t cont
 
 bool SRV_PVDDMON_CheckWindow(void)
 {
-    uint32_t adcValue;
+    uint16_t adcValue;
 
     adcValue = ADC_ConversionResultGet();
     while(adcValue == 0U)
